Favour throwing error on stale params
The idea of these options is that their params will never go stale. They
are initialised with everything which will persist during the life of
the app. If that is not the case then something unexpected has happened
and I would like to make sure I know about it.

I appreciate that this is putting the users at risk of experiencing a
full error rather than any kind of graceful failure which might be the
case with stale references. I'm accepting that cost on the basis that
it's a closed system of limited scope and I think it's most likely that
I will use all of the relevant states to discover issues here if any do
exist.