Replace commonly repeated code with util
I don't understand why this has updated the snapshots. There seems to be
some pattern whereby the sizes have halved. I suspect this is something
which only effects the unit tests rather than the running app.
Unfortunately I won't be able to verify that until I put this code to
the machine I can run the app from. Fortunately this is a speculative
subbranch so if I've made a terrible mistake here I can either fix it or
discard this change.