Remove require cycle
Another approach, and something which I need to get to at some point
would have been to move the styles utils function to the utils folder,
but this is also acceptable. Personally I don't know what is being
referred to by width and height in as intuative a way as if it were
described as short and long, so this is just a helpful wrapper which
other people would also find useful I'm sure. That's why it can
reasonably belong in the `packages` folder.