Repalce hook with improvement
This first commit only looks to drop the improved function in place over
the original. It's no longer a hook so it belongs in a `utils` folder
and it could benefit from having some more tests since it's
functionality is actually improved (see how it actually observes the
`DisplayMode` unlike it's predecessor).

Those actions will wait until a subsequent commit. I wanted to use this
commit to point out how similar they are in concept by keeping the delta
to a minimum.