Remove all legacy 'option' related functionality
An unexpected forced removal but one I'm willing to stand by in the
short term is the ability to change the pozition of the harpstrata by
long pressing on the harpcell. This is an undocumented feature at this
point and I think it can be implemented better so I don't mind tearing
it out for now.

This includes global state keeping track of which of the covariants it
locked and all of the mock objects which had to mock that property.

It also includes all of the hooks associated with selecting values in
the covariant menu. This will expand again as I replace that
functionality with the option-stack implementation, but I think it looks
like it's going to be much lighter weight.