Burst out the option-stack constituent components
This is the first commit in this effort. You can see that we've had to
bring the types and styles up to the src level so there's a bit less
encapsulation of functionality now, but perhaps we can work back towards
it. The truth is that all these components are highly interdependent in
ways which is challenging to capture without repetition but with
localisation. Let's see.