Remove unnecessary shared functions
With the migration of some of the option functions up to the shared
resource level it seems like a good idea to thin out some of the other
stuff that isn't needed. In the process of removing some of these I've
been forced to also update and improve some functionality elsewhere.
This commit might seem a bit offroad, but I think it's valuable in the
context of this feature branch.