Modify option-stack memo areEqual for item equality
This is actually not going to work as a perm solution because while we
do want a rerender of the list objects when the items change, we
actually still don't want the stack to rerender, because as before we
end up resetting the animation value. The only time we should rerender
the option-stack and therefore the animation value is when the number or
options in the stack changes, which should be never.