Reduce memo coverage and allow list rerender in isolation
As discussed in previous commits, we don't want the option-stack to
rerender when the contents of the lists change because that will reset
the animation value to 0 and result in the view snapping back to the
first layer in the stack and scrolling all of the lists back to the 0
index.

Providing the lists with hook derived contents which can be wrapped in a
callback means that none of the option-stack params will go stale by
preventing rerenders, and  it's lists will still be furnished with
updates to their contents.

This is the same principle as we've used with the subtitles which
display the selected values with the `useSubTitle` hook.