Make selected items visbly different and disable touch
This is a POC commit and can probably be optimised by moving the
component creation out to the items creation stage. So each item is just
a component rather than the properties required to build the component.

This commit is acceptable though and is an improvement.

I have had to do some tinkering to the styles to prove the concept and
things are a little out of alignment now, but these can be easily
brought in to shape in the coming commits depending on how the design
progresses.