Fix overlooked mistake in dispatcher config
This bug was introduced in d047e0c7725f72ae04e559045af9466a4297ce95

Evidence, if it were needed, that my test and type coverage is not
perfect.

I'm surprised about the lack of type safety here. See the `TODO` added
which poses what should be an achievable challenge to me and the type
engine.