Build items ahead of list to simplify option types
This has had a great effect on the option types. Because the generic
part is all handled within the item building we don't need it to be
handled further down the chain meaning that a lot of the branching paths
have been removed. See option-list-stack for example.

There is duplication of the `ListItem` component which needs to be spun
out to it's own component, which is why we have a new but currently
unused typeguard which will only become relevant once that has been
done.