Enhance column label stylability
I'm searching for some styling which makes the covariane menu as clearly
readable whilst still conveying *all* of this information. It's proving
difficult, but this is another step forward.

It looks like we're headed for the column labels and the actual menu
titles being basically the same component, especially if the title looks
good with a similar highlight running through it. If that is the case
then they could perhaps be instances of the same component.