Reorient OptionTitle's horizontally
This is better all round, but particularly for the covariant menu.

The semantics of the variables in `getOptionStyles` has been thrown out
a bit, but that can be fixed in upcoming commits.

There is an issue with the natural minor scale name which becomes too
long when turned bold when selected an has to wrap the line. That would
be fine except I want the wrapped line to still justify left. Perhaps an
overflow would be even better.