Simplify option-item in prep for a superscript addition
We need to be able to present the `th` and `rd` suffixes of the
pozitions as we'll as the 'b' for flat suffixes of pitchs in
superscript. We could look at a design which would permit something like
what we have with rendered-tone, or even consider reusing that component
here, but let's take baby steps. I'm not even sure that would look
acceptable in the covariant menu.