Trial a superscript view on position and pitch labels
This is incomplete as it doesn't work well with the positions 10+ (it
superscripts the second digit too).

It also fails to acknowledge that we probably want the same presentation
in the subtitles of the option labels too, and some of the naming is
geting confusing now that several things are called labels. I will
finish the implementation and return to the variable names.