Ensure coherence in option style with variables
These names have still not been audited for their semantics fully. I
need to sweep all of the components later and make sure the naming is
tight, but for now at least this helps make sure that there is coherence
in the sizes chosen and thier relationship to each other.