Use option-text in all option text locations
There was a lot of tinkering required to get this to work. It seems like
the lineHeight which was previously being used to space the items in the
list vertically (which needed to be replcaed because the other contexts
can't have that massive headroom), couldn't be easily replaced by a
height attribute of the same value, because height and lineHeight
interpret the same number differently. I'm sure there's a reasonably
explanation, but for now I've had to settle on fudging the hight a
little bit. None of the stock sizes fit the situation well. In truth
they all made the pink highlight strip end up in unattractive places. I
need to try and fold the highligh strip in to the option-text component
and then try to tidy up.