Localise all the option type definitions
I've checked that this doesn't produce any cyclic dependencies. I'm a
little surprised it doesn't but I'm pleased.

I've had to combine the pointer components as part of this work since I
would otherwise have been left with it's dependency hook needing to
exist all the way in the root hooks which would itself drive the need
for some types to be up there too.