Implement config driven presentation value determination
This still isn't perfect in many ways, but it is definitely better to
have the core objects giving the clues about how their presentational
elements are constructed.

One thing to try to wrestle round in to harmony would be this and the
pitch rendering properties which are used on the harpface.