Change type of iterator used on array
I actually went the opposite direction from what I had suggested in the
TODO. The reason for this is that during normal operation, every item is
going to need to get checked anyway. It's only when something has gone
wrong that the use of `some` would show some performance benefit. At
that point it's not a priority at all. There are faster ways to iterate
anyway, but the good thing about these functions is that they are
semantic. To make the intension as clear as possible here it's better to
use `every`.