Pass in the `useGlobal` dependency for clarity
This clarifies usage a little because now it's obvious that the
callbacks being generated in the menu are defining all of the
dependencies in the dependency list. We define all hooks as being
parameterised.

This isn't complete yet though because there is still a hook being
called from inside this dependency. The useDispatch can be pulled out
and passed in in the same way.