Use window dimension observation hook on problem components
The "problem" components are just the ones which don't have any hooks in
them to resize themselves automatically when the initial harpstrata is
loaded.

However, given that there is no inital harpstrata loaded at all on the
2nd and 3rd pages, are we being too selective here?

Another solution might be to use the `useWindowDimensions` hook at the
very root of the project, but this might overlook some memoised
components.

Another again is to turn the custom size util in to a hook which would
enable all of the sizes to always be completely up to date. The danger
with this though is that I think I've seen this have a huge impact on
initial load time.