Prepare for new tag
What's this new tag about? Nothings changed. That's sort of true. The
build of v8.0.0 was rejected by google because the background location
permissions were still being requested. They shouldn't be since I
upgraded expo to sdk 40: https://dev.to/expo/expo-sdk-40-is-now-available-1in0

I'm hoping that there was just some manifest or other config file which
sill had the location permissions in it and hadn't been cleared out
properly or something, so I've cloned this project fresh and rebuilt it.

Let's see.