Reduce size of close button, prevent overlap
I don't understand where this issue has emerged from, possibly the
upgrade of expo.

Regardles of origin, the styling on this component was already
suboptimual so this is an improvement in all respects.

If you are testing this commit on iOS you will observe that it is not
possible to get to the menus to test this. What's required in the
background (which I haven't committed here because it's under review in
a different pull request) is the iOS quirk workaround I've described in
this forum comment: https://stackoverflow.com/questions/64763186/react-native-reanimated-animation-not-working-on-ios-works-in-android/66143654#66143654