Move option-list from page when inactive
In android it is sufficient to just make the option-list transparent,
then the touches pass through unimpeded. However on iOS that doesn't
appear to be the case. Consequently I need to move the option-lists
aside when they are not being used or they will invisibly block touches
to the option-lists below them.

I have chosen to include an animation to show that one menu is being
swapped out for another which takes advantage of the fact that the
menu's are travelling along the X axis. I'm not certain it's better than
the original fade in and out, but I do like it and the fade in and out
can still be achieved using a condition like the following:

```
const translateX = cond(eq(transitionValue, 0), longEdge, 0)
```

If you are interrogating this commit and are experiencing problems in
iOS then look at the previous commit message for an explanation of why.