Remove SafeAreaView from menu lists
If we're going to make use of SafeAreaView then it needs to go around
the entire app, and the sizes calculated for the entire layout need to
be based on the safe area, not the window size.

Having the SafeAreaView around the list here has meant that while the
rest of the app is in proportion and in the right location, the list is
put out of whack by it's automatic padding which isn't factored in to
the rest of the layout.