Alter option item styling for multi-select lists
The items need to be touchable even when they are selected (so that they
can be deselected) which is not true of single-select lists.

It also won't look right if the selected items are larger than the
deselected ones in this context so the sizing is also made dependent on
this isMultiSelect` property.