Prepare for new tag for iOS re-release
App store connect requires that you submit a new version of the app in
order to update the example images.

I thought I could just submit a request with the same build and
increment the patch id in app store connect only, but it seems I need
a new build too.

Having already created an 8.4.1 release in the app store connect, I am
now forced to release an 8.4.1 version which is unfortunate because this
breaks my compatability versioning standard of not using patch versions.
However, as a one off with an explanation I am willing to accept it.