Add Woozle minor to Woozle Effects own documentation
Woozle Effect seems to include *all* potential overblows in his diagrams
but mark certain ones faintly. I suspect that is to imply that they are
unnecessary because there is a simpler alternatice nearby.

I think perhaps he's done this annotation incorrectly on his diagrams so
in the next commit I'll modify the layout to exclude those which *I*
think need excluding and include only the ones which don't have a
simpler alternative nearby.