Permit negative halfstep index values
With the introduction of half-valved layouts we require some holes
pitches to be lower than the lowest tonic in the key of the harp. This
requires either negative numbers or an index starting from 11. I prefer
to permit negative numbers as it preserves the semantics of there being
a foundational root note, the key of the harp, from which everything
else is derived.