Prevent os font scaling effecting menu layouts
The menu text is big enough for anyone. There is no need for it to scale
as the users os font settings are increased.

It is important that the menu fonts *are not* sized to any other scale
than the one planned since the layout and usability of the menu's is
dependent of the precise sizing of the text.

The fonts elsewhere can be scaled without harm to the layout. Arguably
some of them could have their os scaling removed for some less
significant layout benefit, but this may come at the cost of
readability to users who have chosen large fonts for a reason.