Add test for no overblows on identical reeds
The test scenarios for overblows and bends are slightly different and
I'm not entirely sure I'm comfortable with it.

It does seem however that there are different conventions for when to
acknowledge bends and overblows. And probably there is a different
physics effecting whether one type is more available than another.

For example, there is a bend on the 5 draw, but it's not considered a
full halfstep so no-one acknoledges it on diagrams. I think it's
actually not possible to force it a whole halfstep, unlike the overblows
which are technically always available (at least when there is some
distance between the reeds tuning), even if you don't want to show them
all the time.