Add maaping function for interaction matrix
This is mostly largely based on the halfstepIndexRowMapCallback function
and it's tests but I've pushed the tests a bit further in to the
ridiculous extremes of bending, just to prove that the mapping for each
of the available interaction ids works.

There's no accomodation of the ValvedBend1 interaction id yet since that
functionality is still forthcoming.