Correct type of function return
I realised this while preparing to string the functions together when I
saw that the interaction matrix I was aiming for should actually have
Interaction objects in it rather than InteractionIds.

I've created some techdebt here, captured in the TODO which is that I
have broken my import strategy rather than rewriting the index.ts of the
`constants` folder. I'll pay this off soon.

I am also in danger of creating a cyclic references if the Interaction
constants remain defined in the `constants` folder. I might have to push
them up a level so that both the `utils` and the `constants` folder have
access to them without cyclic references.