Move interactions up to prevent cyclic references
The utils and the constants folder of the apparatus folder are going to
both need to reference interactions and since the constants folder will
reference utils there will be a cyclic reference there unless we take
this action.

Besides, it's not much additional clutter and may even make general
semantic sense.