Change return type to permit caller to report issues
The caller needs to be able to say what was wrong with the hole. It's
important for all the issues with the hole to be interrogated before any
error are thrown so we might as well just return the report from this
function and let the caller do what it wants with it.