Add check to prevent descending bends
All the bends on the holes should be in ascending order and are then
orientated correctly in their matrices based on their role. It's
important that they are ordered ascending at this stage, regardless of
whether the bend goes down like a bend or blowbend, or up like a
overblow or overdraw.