Throw error when producing hole array with errors
The test here probably isn't ideal since it will start to fail when we
fix the hole array generation to not produce this error but rather
filter out the bends in accordance with the physics.

Possibly we should be mocking the component functions so that this is
less an integration test of the error message generation and more a unit
test of the fact that an error is thrown successfully when intended and
it is formatted in a readable way.

This will have to come in a followup commit.