Improve hole validation error message to report apparatus
This felt like modifying the functionality to fit the test initially,
but actually it's modifying the functionality to fit the error message
requirements which is just another part of the functionality, so I'm
justifying this move.

It's a little ugly to have to pass the apparatus as a param, but I don't
see a better way to do it and it's typed so that the develper will be
prompted for it and it's clear that it should be identical to the id of
the apparatus being set as we can see here in this commit.