Split out tuning from apparatus
The idea of this is to eventually remove the constants from apparatus
and turn apparatus in to a module which *builds* apparatus' based on
underlying tuning and some other factors including whether it is valved.

The 2 existing valved tunings have been commented out but not deleted
because the matrices will be useful in the reed-array-to-matrices test
function when we eventually get a way to generate matrices for valved
setups. See the previous paragraph.