Rename prop in HarpStrataProps for accuracy
We're in the middle of really separating the concepts of tuning and
apparatus. The parameter being passed here always was a tuning and
because the apparatus part of the harpstrata didn't have anthing more to
it than tuning, the terms have been synonomous. However, with the
upcoming introduction of valving, we will have more than one thing
comprising the definition of a complete apparatus, so it's important to
increase the accuracy of this parameters naming.