Export renamed function with new name
See previous commits for similar situations with explanations.

This function deserved the rename since the acquisition of an apparatus
is no longer simply a mapped retrieval but rather a process of building.
When we add valving options it will be even more obvious that we're
doing more than we're doing with all the harppart getters which we want
this to be differentiated from.