Rename function for accuracy
In keeping with the other recent commits, we've drawn a clear
distinction between apparatus and tuning. The apparatus will be updated
by this action, but so will it be updated by modifications to the
valving and potentially other things in the future.

We can expect another similar dispatcher to be created when we get
valving added which will presumably be called
`getNewHarpStrataByValvingForDispatcher`.