Handle valved setups in reed-array generation
The `isHalfValved` property is optional at the moment to make this
commit simpler but we will need to make it mandatory shortly.

This commit also sees the known matrices for the valved apparatus which
we were providing in the old setup copied in to the test file to prove
that we're producing the same apparatus as we had before.

These tests aren't quite enough though because it's not testing the
valved draw bends yet.