Rename property for consistency
The other bend types are specific about whether they are blow or draw.
Just because drawbends are the most common and have earned the simple
name doesn't mean it's helpful here. There is actually more work to do
for the names of the interaction constants for example but this is a
good first step.

I've also moved some orderings around in a few places like we tried to
cover off in the previous commit 753afb9, but obviously weren't
complete.