Add valving concept to Apparatus type
This property may need to be made more sophisticated depending on how we
choose to represent valving, but this will do for the moment to enable
us to make a menu which can switch the valving one way or the other.

We're also choosing to leave the optional isHalfValved property  in
`reedArrayToMatrices` as well as having an optional parameter here. Both
of these should be made mandatory in the end, but not yet. The focus for
us now is to get the extra menu working.