Add menu item for choosing apparatus valving
A little reworking of the existing menu has had to be done here to make
sure the naming is consistent. Otherwise this has mostly just been a
case of adding text values to the enum to list as options and copying
the strategy of the other options but to update the harpstrata by
valving this time.