Create ordered valving list
This has mostly been made for consistency with the other harpparts. If
there is a need to have an ordered list of parts in the harpguru-core
project then it makes sense for that ordered list to be defined in the
harpparts project just like all the others.

Because of the standard ordering methodology in harpparts this has meant
that we've now got a (currently) pointless `Valving` type definition and
list of constants, but that's a small price to pay for the consistency
we can benefit from here.

It also means that the source of the valving id's in the harpguru-core
project's valving menu is identical to those being tested in the
apparatus-integration test, also updated here. That's important because
it guarantees that there will be no apparatus' available in the menu
which haven't been tested in this integration test.