Make valvebend mapping function valve-aware
You can see from the detail of this commit, we don't want the logic of
whether to add valvebends outside of the valvebend mapping function.
Similar awareness should be added to the other mapping functions and a
bit of the filtering logic can be removed. We will still need the
validation at the end to make sure that everything has coordinated
effectively.