Add new tunings config and rename awful mappings
I haven't been extremely careful with these configurations because it
will be easiest to verify that the output of the running app which I'll
do shortly.

The reason the renamed mappings were "awful" is because they should have
been caught in the code review of the pull request which introduced
them.