Rename Richter tuning for presentational reasons
I would have liked to have the "major diatonic" in there for people who
call it that and might not know it by "Richter", but I think most people
do, and those who don't probably won't be playing with tunings anyway.

The layout was too cramped with the "major diatonic" included and that
was a bad look for it considering it was the default option selected.