Rename Richter tunings appropriately
The plan is to have a complete list of the Richter tunings under a
Richter heading so they don't need to have the word Richter in their
name.

I have chosen Aeolian rather than Natural Minor to differentiate it from
the tuning commonly known as Natural Minor, which is really a Dorian
Richter tuning considered from a second position perspective.

This list is probably mostly for accademic purposes so I think it's ok
to use the mode names without any concern for how that might commonly be
perceived.

I will probably want to add Ionian in to this too.