Add full set of spiral tunings
Having to have the "(spiral)" in there is a shame but they need to be
differentiated from richter enumerators. If I want to get rid of this I
can just have a name property on the tuning objects which is free text
and the enumerators can be named whatever they need to be. Possibly I'll
not want to return them to numeric enums for debugging purposes.