Tidy tuning constants id definitions
There are two ideas here. One is that we *know* that these are tuning
id's because it is the id property of the tuning objects so there's no
need for each line to have the `TuningIds` prefix clarifying what they
are. The second is that it will be easier to check that all the tuning
ids have been imported if they are all explicitly stated together at the
top of the file.