Replace useSizes hook in to direct component locations
It will take a little more work to replace the getSizes util entirely
since it's being used in some style utils which themselves are not
called from hook complient contexts.

This will serve as a good demonstration of how this might function,
although a full transplant will eventually be necessary to get a full
impression.