Replace useSizes hook in to options sizing
This is definitely a step too far. Potentially the correct way to handle
this is to continue using this useSizes hook, but to give it some static
values too which can be used for the parts of the app which we don't
want to resize with the harpface. I think I favour this to having a
separate getSizes function which will otherwise mean redundancy  which I
don't think is necessary or helpful here.