Add sliceMatrix function
This function is a candidate for truncating the ends off of harpstrata
for the purpose of presenting smaller fragments as "zooms" on the
harpface.

I want it to behave similarly to the array `slice` function which is why
I'm messing around with these defaulting values where I probably don't
need them.