Introduce column bounds via getHarpFaceFacts
This has proven to be a bit messier than I had hoped. The main problem
is actually that the getHarpFaceFacts and associated styles are
requiring more information thatn they really should need and binding
things together as a result.

I think I might try a new branch where I simplify the getHarpFaceFacts
by moving some of the information they require in to the global state
before trying this again.