Move octave group flattening in to hook
This tidies up the component a bit and makes this extra bit of logic
testable too.

We have also improved the functionality a bit because we can see here
that the snapshot was previously not actually setting a different width
on the harpface when the fragmentation was on vs off. No visual impact
in the app during normal use, but there would be some edge cases where
this might have had an impact. The reason was that the useStyles hook
wasn't flattening things out the same way as the component so they were
slightly out of whack in their assumptions.