Do not render empty rows resulting from column bounding
When we bound the columns which are rendered, we often end up with rows
which no longer have a cell for rendering. Often a row only has one cell
on it (the 3 triple draw bend for example, or the 10 double blowbend).
When we bound those columns out of the picture, we need to prevent the
rendering of that row so we don't waste space on the page.