Make layout sensitive to viewable window
Even though the correct number of cells was being displayed we were
actually still basing the layout logic on the full harp so things were
not presenting correctly. This is why one of the snapshots has been
updated in this commit.

I have also been lazy with the import of the useViewableMatrices
function. It's still sitting in the HarpFaceFragment folder which is now
inappropriate given that it's being used for the layout too. I'll move
it in the next commit.