Resolve require cycle
The refactoring here is in keeping with the strover project formatting
I've been trying to adhere to later in the life of the project. The
src level `styles` folder was never in keeping with that and it's good
that I've had the opportunity to attack it. I'll continue in the next
commit even though it's not really part of the feature being worked on
here.