Move color logic to utils folder
This follows on from the movement of the sizing logic to the hooks
folder.

The purpose of this is simply to have the color logic fit the project
folder strategy better than it did before. It's not necessitated by
the feature under development, but it did seem sensible to do it
since the sizes had the same problem and they have now been improved.