Create stepping stone zomm logic util
The logic which is currently in `get-new-column-bounds-for-dispatcher`
was just jotted out to test a UI idea I had. It's time to be a little
more methodical about it. I'm also pretty sure that this logic is going
to be needed not only in that menu, but also at a page level to deal
with the updating of activeHarpStrata to those with different lengths.
Here's the start of a more careful step in to the unknown.