Add test for best 7 holes not dropping below 0
I'm still not sure whether this should perhaps be trimming the overflow
off or replacing with 'FIT' when there's actually fewer harp holes than
the bounds are restricting. This might mess up the face column
estimates. I'm going to keep it simple for now though and see what
happens.

The benefit of doing it like this is that uses don't lose their zoom
when they accidentally select a harp which is already less than that
zoom.