Add util to assist columnBounds dispatcher
See in the upcoming commits, this will help set a default value for the
inbound zoomId. The idea is that if the zoomId property of the
dispatcher is not set then we can determine what it's value was last
time it was set by inferring it from the columnBounds.

We don't want to have to store the zoomId *and* the columnBounds in
global scope since the information would be redundant.