Make dispatcher multifunctional
This dispatcher can now work from the UI context where the
activeHarpStrata is static and the zoomId is driving the modification of
the columnBounds *and* from the context of the activeHarpStrata being
updated and the columnBounds needing to adaptively respond.

This dispatcher will now be moved to a location where it can be put to
both these uses.