Add component to review columnBounds at harpstrata update
An interesting visual quirk is occuring with this component operating
this way. What seems to be happening is that certain cells are being
rerendered, but the activity status of them isn't updated until a
subsequent render. That breaks the continuity between one harp and the
next and it's difficult to compare them.

Possibly the problem is that we shouldn't be waiting until after the
activeHarpStrata has updated before taking action on the columnBounds.
useEffect might not be the best way to respond to that act since I think
it's asynchronous. Perhaps there is a more direct way of responding to
the update of activeHarpStrata.