Prepare for alternate columnBounds update solution
I'm struggling to see why I'm getting the delayed update issue when
using the `useUpdateColumnBoundsByActiveHarpStrata` hook, and I'm
struggling to find ways to fix it.

In this commit I'm laying the foundations of an alternative.

I have created a `determineNextColumnBounds` function which will most
likely stay even if the new hook doesn't. It needs to be used in the
`getNewColumnBoundsForDispatcher` if that stays.