Collect elements of columnBounds update together
The elements of the columnBounds update have become quite disorganised.
Here I pull them all together again. Probably not in their final
location, but somewhere they can be made more sense of at least.

The dispatch callback is now in the root utils folder because currently
it needs to be called by the menu-of-tunings when the zoom options are
modified, and also by the SyncColumnBoundsWithActiveHarpStrata
compopnent.

I think that I will have to replace the
SyncColumnBoundsWithActiveHarpStrata component with some other solution
since I'm still unable to resolve the render interruption which appears
to take place before the active cells are rendered. If I do replace it
then all of these files might need to move again, but at least I've got
the utility files together an not scattered all over the place now.