Add global reducers to bind columnBounds to harpStrata
This is a partial commit. Functionally it seems to work if you make the
relevant commented out changes to
`get-new-harpstrata-by-covariants-for-dispatcher` but from a typing
perspective there are still issues to resolve which I'll try to deal
with in the next commit.

I also need to actually put the logic in the `nextColumnBounds` reducer
and fully remove the `sync-column-bounds-with-active-harp-strata`
component.