Replace global reducers with local ones
The global reducer idea would have actually been very appropriate in
this context. Have a global reducer sat in front of all the global state
to make it clear that no global state should be updated directly. As it
stands there is a referencial cycle which can't be avoided when using
typescripts `addReducer` function which reactn uses for it's own
`addReducer`: https://github.com/CharlesStover/reactn/issues/174

Charles's own advice is to use local reducers instead. I still want to
get some kind of benefit like that described above, and while I don't
think it's actually as semantically clear doing it locally like this
there probably is scope to have saga's like reducers like the one I've
created here at `setNewHarpStrataReducer`. Making this pattern
ubiquitous in the way required though would be out of scope for this
feature. I'll have to raise it as techdebt.