Remove redundant TODO
It might have been better to name the variable `holeBounds`, but
`columnBounds` is actually fine and there are 155 instances of it in the
code now so updating it will not be worth the effort.

The second point is actually not valid now either. I made a decision
with the UI component to zoom levels of more than the number of holes
that the harp has so that the user can lock in a specific zoom level
even while changing around the harps.