Make reducer match other reducer standards
I'm trying to create dispatch sagas here, without the use of the
`dispatch` param because typescript complains about cyclic references.
Consequently, a pattern I'm establishing now is that all the reducers
will receive the `global` parameter as it's starting value at the
beginning of the update.

I may regret this decision since it's not the way that the `global`
property is passed around during proper sagas, but trying to reproduce
that approach by hand makes the saga's less easy to read I think, which
is important.