Ensure original state is returned if matches new
It may still be better yet to simply not return anything rather than an
identical object. I am just going by the documentation at the moment.
Once I can see the benefits in the running app I will swap it out for
that functionality and see if things continue to work as hoped.