Lay template for replacement new harpstrata reducers
The strategy described in the comment in the previous commit may have to
be started prematurely for me to even continue to make sense of what I'm
doing. I am still skimping on the unit tests and separation of utils for
the moment though, hense all the TODO comments.

This is not even a complete template yet. This is just a replacement of
the reducers that have been hacked together so far.