Add layoutFacts draft reducer
This is the last piece of the exploded global state that we need for the
experiment I think.

I may not jump straight in to those experiments since I'm sat in front
of the tv right now and can't actually run the app to test the changes.
I may start by beginning to tidy up and add tests to the composite
reducers.